refactor: split buffers/sizes.rs and qwen3_attention/types.rs under the 500-LoC cap
Merging main (#281's buffer-split + FP4/SSD additions) with this branch's
LoRA buffer fields pushed two files over the 500-line cap:
  buffers/sizes.rs        503 -> 464  (+ sizes_q12.rs: the Q12 batched-prefill
                                       scratch const + helper)
  qwen3_attention/types.rs 504 -> 335  (+ types_weights.rs: MLA / compressor /
                                       hyper-connection weight structs)
Exact piecewise move, re-exports preserved. Builds clean; fmt/clippy/test/docs
green; 0 cap violations.
